# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "File Conversion Platform",
    "description": "Comprehensive file conversion platform supporting 100+ conversion types",
    "version": "1.0.0"
  },
  "paths": {
    "/api/documents/convert/{conversion_id}": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Convert Document",
        "description": "Convert a document to the specified target format\n\n- **conversion_id**: ID of the uploaded file\n- **target_format**: Target format (pdf, docx, txt, csv, xlsx, etc.)",
        "operationId": "convert_document_api_documents_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          },
          {
            "name": "target_format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Target Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/convert/{conversion_id}": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Convert Image",
        "description": "Convert an uploaded image to the specified format.\n\nSupported conversions:\n- JPG Γåö PNG\n- PNG ΓåÆ WebP\n- WebP ΓåÆ PNG/JPG\n- TIFF ΓåÆ JPG/PNG\n- BMP ΓåÆ JPG/PNG\n- HEIC ΓåÆ JPG/PNG\n- SVG ΓåÆ PNG/JPG\n- Image ΓåÆ PDF\n- ICO Γåö PNG",
        "operationId": "convert_image_api_images_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_image_api_images_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/video/convert/{conversion_id}": {
      "post": {
        "tags": [
          "video"
        ],
        "summary": "Convert Video",
        "description": "Convert video files between formats or extract content.\n\nSupported conversions:\n- MP4 Γåö MKV, AVI\n- MOV ΓåÆ MP4\n- FLV ΓåÆ MP4\n- WMV ΓåÆ MP4\n- WebM ΓåÆ MP4\n- Video ΓåÆ GIF\n- Video ΓåÆ Audio\n- Video ΓåÆ Frames",
        "operationId": "convert_video_api_video_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_video_api_video_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/security/convert/{conversion_id}": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "Convert Security",
        "description": "Perform security and encoding conversions.\n\nSupported conversions:\n- File Γåö Base64\n- File ΓåÆ Hash (MD5, SHA256, SHA512)\n- File Γåö Encrypted\n- PDF ΓåÆ Locked PDF\n- Locked PDF ΓåÆ PDF",
        "operationId": "convert_security_api_security_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_security_api_security_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/download/{conversion_id}": {
      "get": {
        "summary": "Download File",
        "description": "Download converted file",
        "operationId": "download_file_api_download__conversion_id__get",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_convert_image_api_images_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_image_api_images_convert__conversion_id__post"
      },
      "Body_convert_security_api_security_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm",
            "default": "sha256"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_security_api_security_convert__conversion_id__post"
      },
      "Body_convert_video_api_video_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_video_api_video_convert__conversion_id__post"
      },
...
      "ConversionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "target_format": {
            "type": "string",
            "title": "Target Format"
          },
          "file_size": {
            "type": "number",
            "title": "File Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "source_format",
          "target_format",
          "file_size",
          "status",
          "created_at"
        ],
        "title": "ConversionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}